Moved handling advertise address to readConfig and out of the agent's constructor, plus unit test fixes
