+++ agent/local_test.go
@@ -14,7 +14,9 @@ import (
 
 func TestAgentAntiEntropy_Services(t *testing.T) {
 	t.Parallel()
-	a := &TestAgent{Name: t.Name(), NoInitialSync: true}
+	cfg := TestConfig()
+	a := &TestAgent{Name: t.Name(), NoInitialSync: true, Config: cfg}
+
 	a.Start()
 	defer a.Shutdown()
 
@@ -670,8 +672,10 @@ func TestAgentAntiEntropy_Services_ACLDeny(t *testing.T) {
 
 func TestAgentAntiEntropy_Checks(t *testing.T) {
 	t.Parallel()
-	a := &TestAgent{Name: t.Name(), NoInitialSync: true}
+	cfg := TestConfig()
+	a := &TestAgent{Name: t.Name(), NoInitialSync: true, Config: cfg}
 	a.Start()
+
 	defer a.Shutdown()
 
 	// Register info
