+++ agent/testagent.go
@@ -334,6 +334,7 @@ func TestConfig() *Config {
 
 	ccfg.CoordinateUpdatePeriod = 100 * time.Millisecond
 	ccfg.ServerHealthInterval = 10 * time.Millisecond
+	cfg.SetupTaggedAndAdvertiseAddrs()
 	return cfg
 }
