+++ command/agent_test.go
@@ -176,6 +176,7 @@ func TestReadCliConfig(t *testing.T) {
 			args: []string{
 				"-data-dir", tmpDir,
 				"-node", `"a"`,
+				"-bind", "1.2.3.4",
 				"-advertise-wan", "1.2.3.4",
 				"-serf-wan-bind", "4.3.2.1",
 				"-serf-lan-bind", "4.3.2.2",
@@ -207,6 +208,7 @@ func TestReadCliConfig(t *testing.T) {
 				"-data-dir", tmpDir,
 				"-node-meta", "somekey:somevalue",
 				"-node-meta", "otherkey:othervalue",
+				"-bind", "1.2.3.4",
 			},
 			ShutdownCh:  shutdownCh,
 			BaseCommand: baseCommand(cli.NewMockUi()),
@@ -229,6 +231,7 @@ func TestReadCliConfig(t *testing.T) {
 				"-node", `"server1"`,
 				"-server",
 				"-data-dir", tmpDir,
+				"-bind", "1.2.3.4",
 			},
 			ShutdownCh:  shutdownCh,
 			BaseCommand: baseCommand(ui),
@@ -256,6 +259,7 @@ func TestReadCliConfig(t *testing.T) {
 			args: []string{
 				"-data-dir", tmpDir,
 				"-node", `"client"`,
+				"-bind", "1.2.3.4",
 			},
 			ShutdownCh:  shutdownCh,
 			BaseCommand: baseCommand(ui),
@@ -301,6 +305,7 @@ func TestAgent_HostBasedIDs(t *testing.T) {
 		cmd := &AgentCommand{
 			args: []string{
 				"-data-dir", tmpDir,
+				"-bind", "127.0.0.1",
 			},
 			BaseCommand: baseCommand(cli.NewMockUi()),
 		}
@@ -317,6 +322,7 @@ func TestAgent_HostBasedIDs(t *testing.T) {
 			args: []string{
 				"-data-dir", tmpDir,
 				"-disable-host-node-id=false",
+				"-bind", "127.0.0.1",
 			},
 			BaseCommand: baseCommand(cli.NewMockUi()),
 		}
